requirements/base.txt
pytz==2019.2  # https://github.com/stub42/pytz
redis==3.3.8  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
flower==0.9.3  # https://github.com/mher/flower
kombu==4.6.3
django-crispy-forms==1.9.0

# Django
# ------------------------------------------------------------------------------
django==2.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-redis==4.10.0  # https://github.com/niwinz/django-redis

-e git+https://github.com/kamakazikamikaze/easysnmp.git@refs/pull/106/merge#egg=easysnmp
